Introduce SquareGridGroupCover ans Refactor PlaylistGroupCover
